refactor: switch to interface rather than impl for Ignition object
Switch the Ignition object in the type extension of NetworkConnection to
the interface rather than the impl, so the impl can move to internal.
